requirements.txt
tensorflow==2.15.0
numpy>=1.19.2
opencv-python>=4.5.1
PyYAML>=5.4.1
scikit-learn>=0.24.2
wandb>=0.12.0
matplotlib>=3.3.4
tqdm>=4.61.0 